Add testDeleteIDNotFound test class

+++ src/main/java/muslimdev/spring6restmvc/Services/BeerServiceJPA.java
@@ -60,10 +60,15 @@ public class BeerServiceJPA implements BeerService {
     }
 
     @Override
-    public void deleteById(UUID beerId) {
+    public Boolean deleteById(UUID beerId) {
 
-        beerRepository.deleteById(beerId);
+        if (beerRepository.existsById(beerId)){
 
+            beerRepository.deleteById(beerId);
+            return true;
+        }
+
+        return false;
     }
 
     @Override
